Synchronized docker wrapper around build process
Fixes #81.  There were a few minor things that had gotten out of sync.
First was updating Golang to match the makefile expectations.
Second was to add a step to acquire go deps to the makefile.

I also took the liberty to add support for making the release artifacts
available outside docker as well.  Not strictly required, but seemed
nice to have.